requirements.txt
certifi==2020.12.5
chardet==4.0.0
debtcollector==2.2.0
dnspython==1.16.0
eventlet==0.30.0
greenlet==1.0.0
idna==2.10
importlib-metadata==3.4.0
msgpack==1.0.2
netaddr==0.8.0
oslo.config==8.4.0
oslo.i18n==5.0.1
ovs==2.13.0
pbr==5.5.1
PyYAML==5.3.1
repoze.lru==0.7
requests==2.25.1
rfc3986==1.4.0
Routes==2.5.1
ryu==4.34
scapy==2.4.4
six==1.15.0
sortedcontainers==2.3.0
stevedore==3.3.0
tinyrpc==1.0.4
typing-extensions==3.7.4.3
urllib3==1.26.2
WebOb==1.8.6
wrapt==1.12.1
zipp==3.4.0
